requirements.txt
allure-pytest==2.8.40
allure-python-commons==2.8.40
apipkg==1.5
Appium-Python-Client==2.6.0
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
cffi==1.15.1
chardet==4.0.0
chromedriver-autoinstaller==0.3.1
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cryptography==37.0.4
decorator==5.1.0
enum34==1.1.10
et-xmlfile==1.1.0
execnet==1.8.0
h11==0.13.0
idna==2.10
iniconfig==1.1.1
lxml==4.6.3
msedge-selenium-tools==3.141.3
namedlist==1.8
openpyxl==3.0.7
outcome==1.2.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.2
pytest-excel==1.4.2
pytest-forked==1.3.0
pytest-rerunfailures==9.1.1
pytest-runner==6.0.0
pytest-variables==2.0.0
requests==2.25.1
selenium==4.3.0
self==2020.12.3
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
toml==0.10.2
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.4
webdriver-manager==3.4.0
wsproto==1.1.0
